Simplify crawler scripts and handle missing data
