--- crawler/uploader.py
@@ -1,13 +1,0 @@
-# crawler/uploader.py
-# In this GitHub Actions setup, the uploader's role is performed by git commands
-# in the workflow YAML file. This file is kept for structural completeness and
-# could be used for more complex upload logic in the future (e.g., uploading
-# to a different storage service).
-
-def upload_results():
-    """
-    Placeholder for uploading results. The actual upload is done via git
-    commands in the .github/workflows/crawl.yml file.
-    """
-    print("Upload is handled by the GitHub Actions workflow.")
-    pass
